+++ src/main/java/com/bad/bigad/service/PlayerService.java
@@ -3,7 +3,7 @@ package com.bad.bigad.service;
 import com.bad.bigad.entity.Player;
 
 public interface PlayerService {
-    public Player findById(int id);
+    public Player findById(long id);
     public Player findByWxName(String wx_name);
     public Player CreateNew(String wx_name, String wx_nick_name);
 }
